pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "hpa-mdo"
version = "0.1.0"
description = "Human-Powered Aircraft Multidisciplinary Design Optimization Framework"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "numpy>=1.24",
    "scipy>=1.10",
    "pandas>=2.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "matplotlib>=3.7",
]

[project.optional-dependencies]
oas = ["openaerostruct>=2.7", "openmdao>=3.28"]
api = ["fastapi>=0.100", "uvicorn>=0.23"]
mcp = ["mcp>=1.0"]
dev = ["pytest>=7.0", "pytest-cov", "ruff"]
all = ["hpa-mdo[oas,api,mcp,dev]"]

[project.scripts]
hpa-mdo = "hpa_mdo.api.server:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"
